renamed/refactored/cleaned-up TilesetOptions
